Update StorageUnits.cs
[Sorry, I couldn't resist :P](https://en.wikipedia.org/wiki/Binary_prefix)

On a more serious note, I'm fine with either of the nomenclatures, but mixing them is not a good idea. And by that I mean that you filled in 1000 as the kilobyte limit, but used 1024 everywhere else, including where the actual counting happened. This resulted in 1000 bytes being printed out as "0.97 KB", which I'm fairly sure was not the design goal, even if you prefer KB' over KiB'.

I'm also not terribly sure why haven't you just casted the limits to doubles and used them for the divisions, or better yet, just made them doubles to begin with, but what do I know ¯\_(ツ)_/¯. I did correct all the output strings to use binary prefixes though, but if you don't like them, I'll revert it gladly.

:popcorn: